Use favorite hook exported outside video card
